feat: --emit flash + AVR format fixes + elf cleanup
- New --emit flash mode: tsc→c→avr-gcc→elf→hex→avrdude→device
- Flash config in tsc.package.json (programmer/port/baud/extraFlags)
- Early validation: target check + flash config check
- .elf cleanup on avr-gcc/objcopy failure
- _tsc_format_impl custom parser for AVR (replaces %lld/%zu/%.*s)
- Embedded format fixes: %ld/%lu for i32/u32, manual i64/u64 conversion
- 2 phase11 expected.c fixes (%zu→%u, %d→%ld)
- 3 new flash error tests + 2 AVR template literal tests
- Spec 13-build.md updated with flash field
- phase9: 25→53 tests

+++ test/cases/phase11/allocator-none/fixed-array-ok/expected.c
@@ -3,6 +3,6 @@
 int main(void) {
     TSC_INIT();
     uint8_t buf[256] = {0};
-    printf("%zu\n", (size_t)256);
+    printf("%u\n", (unsigned)(size_t)256);
     return 0;
 }
