+++ test/cases/phase11/allocator-none/value-type-stack/expected.c
@@ -5,6 +5,6 @@ typedef struct { int32_t x; int32_t y; } Node;
 int main(void) {
     TSC_INIT();
     Node n = { .x = 0, .y = 0 };
-    printf("%d\n", n.x);
+    printf("%ld\n", (long)n.x);
     return 0;
 }
